pyproject.toml
[tool.poetry]
name = "dpt"
version = "0.1.0"
description = ""
authors = ["Vladimir-Vologda <vologda2018@inbox.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.3"
djangorestframework = "^3.13.1"
Pillow = "^9.0.1"
django-environ = "^0.8.1"
psycopg2-binary = "^2.9.3"
django-phonenumber-field = "^6.1.0"
phonenumbers = "^8.12.45"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
